fix #6

+++ hash_table/src/main/HashTable.java
@@ -34,7 +34,7 @@ public class HashTable
         int n = value.length();
         int hash = 0;
         for (int i = 0; i < n; ++i) {
-            hash += powers[i] * (value.charAt(i) + 1);
+            hash += powers[i] * ((int)value.charAt(i) + 1);
             hash += MOD;
             hash %= MOD;
         }
